pyproject.toml
[project]
name = "gliner-api"
version = "0.3.5"
description = "Easily configurable API & frontend providing simple access to dynamic NER models"
readme = "README.md"
requires-python = "==3.13.7"
dependencies = [
    "fastapi[standard]==0.116.1",
    "gitpython==3.1.45",
    "huggingface-hub==0.34.4",
    "prometheus-client==0.22.1",
    "pydantic-settings==2.10.1",
    "uvicorn[standard]==0.35.0",
]

[project.optional-dependencies]
cpu = [
    "torch==2.8.0",
    "gliner[stanza]==0.2.22",
    "onnxruntime==1.22.1",
]
gpu = [
    "torch==2.8.0",
    "gliner[stanza]==0.2.22",
    "onnxruntime-gpu==1.22.0",
]
frontend = [
    "gradio==5.44.1",
    "httpx==0.28.1",
    "stamina==25.1.0",
]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "gpu" },
    ],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu128", extra = "gpu" },
]

[[tool.uv.index]]
name     = "pytorch-cu128"
url      = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name     = "pytorch-cpu"
url      = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = [
    "ipykernel==6.30.1",
    "pre-commit==4.3.0",
    "ruff==0.12.12",
    "ty==0.0.1a20",
]
